requirements.txt
flask
flask-cors
python-dotenv
requests
matplotlib
networkx